requirements.txt
Django==2.1.4
django-widget-tweaks==1.4.2
google-cloud-vision==0.35.1
autocorrect==0.3.0
google-api-core==1.6.0
google-auth==1.6.1
googleapis-common-protos==1.5.5
Pillow==5.3.0
psycopg2-binary==2.7.5
whitenoise==4.1.2